Change files extensions to jsx
